Move secondSign tests to separate folder
