+++ test/functional/system/1.second_signature/2.X.second_signature_unconfirmed.js
@@ -15,11 +15,11 @@
 'use strict';
 
 var lisk = require('lisk-js');
-var accountFixtures = require('../../fixtures/accounts');
-var randomUtil = require('../../common/utils/random');
-var normalizer = require('../../common/utils/normalizer');
-var transactionTypes = require('../../../helpers/transaction_types.js');
-var localCommon = require('./common');
+var accountFixtures = require('../../../fixtures/accounts');
+var randomUtil = require('../../../common/utils/random');
+var normalizer = require('../../../common/utils/normalizer');
+var transactionTypes = require('../../../../helpers/transaction_types.js');
+var localCommon = require('../common');
 
 describe('system test (type 1) - sending transactions on top of unconfirmed second signature', () => {
 	var library;
