+++ test/functional/system/1.second_signature/3.X.second_signature_validated.js
@@ -15,11 +15,11 @@
 'use strict';
 
 var lisk = require('lisk-js');
-var accountFixtures = require('../../fixtures/accounts');
-var randomUtil = require('../../common/utils/random');
-var normalizer = require('../../common/utils/normalizer');
-var transactionTypes = require('../../../helpers/transaction_types.js');
-var localCommon = require('./common');
+var accountFixtures = require('../../../fixtures/accounts');
+var randomUtil = require('../../../common/utils/random');
+var normalizer = require('../../../common/utils/normalizer');
+var transactionTypes = require('../../../../helpers/transaction_types.js');
+var localCommon = require('../common');
 
 describe('system test (type 1) - checking validated second signature registrations against other transaction types', () => {
 	var library;
